+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add .gitignore
 - Add demo for OAuth 2.0 implicit grant type
 - Add demo for OAuth 2.0 password grant type
+- Change load-test-introspect.js to generate access token by code grant instead of client credentials grant.
 
 ## 2024-11-24